Fixed some multiplier variables and added main notebook

+++ train_uniter.py
@@ -577,7 +577,7 @@ if __name__ == '__main__':
                         help='Normalize images by dividing them by their height and width. Default=True')
     parser.add_argument('--train_filename', type=str, default='train.jsonl',
                         help='The name of the trainin json file to load.')
-    parser.add_argument('--upsample_multiplier', type=int, default=0,
+    parser.add_argument('--upsample_multiplier', type=float, default=0,
                         help='Multiplier used to increase the amount of confounders in training data')
     parser.add_argument('--note', type=str, default='',
                         help='Add a note that can be seen in wandb')
